requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
datasets==2.19.2
diffusers==0.28.2
dill==0.3.8
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
huggingface-hub==0.23.3
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
open-clip-torch==2.24.0
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==5.27.1
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
sympy==1.12
timm==1.0.3
tokenizers==0.19.1
torch==2.3.1+cu118
torch-summary==1.4.5
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
